cuda/requirements.txt
cupy-cuda13x==13.6.0
Cython==3.2.2
fastrlock==0.8.3
Mako==1.3.10
MarkupSafe==3.0.3
numpy==2.3.4
platformdirs==4.5.0
pycuda==2025.1.2
python-dotenv==1.2.1
pytools==2025.2.5
siphash24==1.8
typing_extensions==4.15.0
